+++ src/test/java/micycle/pgs/PGSTests.java
@@ -7,8 +7,6 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.fail;
 
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
 import java.util.function.UnaryOperator;
 
@@ -21,39 +19,11 @@ import org.locationtech.jts.geom.LinearRing;
 import org.locationtech.jts.geom.MultiPolygon;
 import org.locationtech.jts.geom.Polygon;
 
-import micycle.pgs.commons.PEdge;
 import processing.core.PConstants;
 import processing.core.PShape;
 import processing.core.PVector;
 
 class PGSTests {
-
-	@Test
-	void testFromEdgesSimple() {
-		PEdge a = new PEdge(0, 0, 1, 1);
-		PEdge b = new PEdge(1, 1, 1, 0);
-		PEdge c = new PEdge(1, 0, 0, 0);
-
-		List<PEdge> edges = Arrays.asList(a, c, b); // a, c, b
-
-		List<PVector> orderedVertices = PGS.fromEdges(edges);
-		assertEquals(3, orderedVertices.size());
-	}
-
-	@Test
-	void testFromEdges() {
-		List<PEdge> edges = new ArrayList<>();
-		for (int i = 0; i < 15; i++) {
-			edges.add(new PEdge(i, i, i + 1, i + 1));
-		}
-		edges.add(new PEdge(15, 15, 0, 0)); // close
-
-		Collections.shuffle(edges);
-
-		List<PVector> orderedVertices = PGS.fromEdges(edges);
-		PGS.fromEdges(edges).forEach(q -> System.out.println(q));
-		assertEquals(16, orderedVertices.size());
-	}
 
 	@Test
 	void testOrientation() {